pyproject.toml
[project]
name = "wordle-llm"
version = "0.1.0"
description = "Wordle LLM"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "wandb",
    "vllm",
    "trl",
    "Pillow",
    "html2image",
    "huggingface_hub[cli]",
    "litellm",
    "python-dotenv",
    # "torch>=2.7.0",
    # "transformers",
    # "accelerate>=1.4.0",
    # "peft",
    # "rich",
    # "trl>=0.17.0",
    # "vllm[flash-attn]>=0.9.2",
    # "liger-kernel>=0.5.10",
    # "deepspeed",
]


[project.optional-dependencies]
dataset = [
    "spacy",
    "inflect",
    "wordfreq",
]